r53spflat/requirements.txt
boto3>=1.23.3
dnspython>=2.2.0
netaddr>=0.8.0
sender-policy-flattener>=0.3.1
slack_sdk
